TBC Time To Expose Simulation
I made this tool back in TBC to compare different opener options. It didn't have a home so I have placed it here for perusal.

This tool mainly allows you setup some opener parameters like whether to cast Adrenaline Rush before expose (to get expose up faster), or to cast Adrenaline Rush after expose (to deal more damage with the Adrenaline Rush with expose on the target). You can also change very many things like Blade Flurry timings, whether you are in front of the mob for a few seconds on the opener and are susceptible to parries, if you have a proc trinket that provides haste (like DST), and a whole slew of other options.

I think it's a really cool tool, and if you are at all familiar with TBC expose openers you might think this is cool too.

It's a python script, and you want to open the code and directly modify the parameters at the top, then once you're done save and run it. Save your results to a notepad, change some settings, and run it again, and compare results to your notepad. Works great in online python compilers, just google "online python". Just follow the instructions in the comments of the code and you should get there in the end.